Added analysis, physics modules, new grid option
Added analysis module for helping with data analysis.

Begun work on the physics module for computing transport physics.

Begun work on the physics_to_H module for converting transport physics
quantities to the H coefficients for the transport equation.

Added a new option to gridinterface_gene for interfacing grids between
gene and tango.  The new option uses a fixed slope when mapping a
proflie from tango to gene.