new derivatives module; compute_all_H now a function
Moved the calculation of derivatives of an array into its own module.
Previously individual modules had their own computations of derivatives.

The compute_all_H argument passed into the solver class is now more
general.  It no longer takes "turbhandler" as an argument.  Instead,
compute_all_H is to be thought of as an *object*, not a basic function.
Hence, it contains a reference to turbhandler (as well as a reference to
any other transport physics one wants to include) without the solver
class having to know about it.  Using some Python magic, the object is
itself callable the way a function would be.  Tests and examples using
the solver class were updated to this new style.