[handlers] New
Added handlers module for dealing with saving files to disk, with the
purpose of ongoing saves admist a Tango run as contingency in case a run
does not finish (a very likely possibility).  The solver can now use
both the datasaver and the handlers.

Three handlers have been created:
1) for saving the distribution function f1 checkpoint history
2) saving a checkpoint for Tango from which Tango can be easily
restarted
3) saving Tango history in an ongoing fashion, for post-run analysis to
debug what went wrong.  This handler works very similarly to datasaver,
but saves to disk in an ongoing basis instead of just at the end of a
run.

The module uses an MPI-based rank to ensure that only one process is
saving the specified data to disk, to ensure that multiple Python
processes in a parallel environment do not all try to write the same
data.

Also updated datasaver module to work in a parallel environment in the
same way.