[lodestro_method] derivative at bndy now 1st-order
In the lodestro_method module, the derivative at boundary points was
changed from 2nd-order to 1st-order finite difference.  It originally
used 1st order, but an Oct. 25, 2016 commit changed it to 2nd order.
This caused a bug, undiscovered until now, where for small values of the
EWMA parameter, the solver was unable to converge to a fully accurate
solution to machine precision.

This went unnoticed because all automated tests used a value of the EWMA
parameter of 0.3.  This bug didn't manifest until a value <~ 0.01 is
used.   A new test was added that checks for convergence down to high
accuracy with this EWMA parameter.