[datasaver] Minor reworking
The datasaver module has been reworked slightly, to allow another method
of interactive access.  Previously, all of the data would be
automatically saved to disk and then wiped from working memory.  Now,
the data is not wiped at the end of the last timestep. This way, if one
does not care for the data to be saved to disk, the data can still be
accessed and interacted with.