[bufferzone] decorator; [interfacegrids] extrapolator
Added a bufferzone module, which serves as a decorator for applying a
GENE-like buffer zone to the flux returned by a FluxModel.

In the interfacegrids_gene module, a new GridMapper class has been
added, which extrapolates the transport coefficients into the buffer
zone (and beyond) using the transport coefficients near the buffer zone,
in what is called the extrapolation zone.  Extrapolation is performed
only on the right boundary of the domain.

An example demonstrating the use of the buffer zone decorator and the
extrapolation, and comparing the result to the Shestakov problem with no
buffer zone, has also been added.