[solver] User control func for customized control
Added user_control_func which can optionally be passed into solver on
construction.  It runs once per iteration.  It allows the user to
customize control over the course of the iteration, using external
knowledge not available to the Tango core infrastructure.  A test has
been added which demonstrates the use.