Added files for creating a flux seed
Added the module read_fluxprof2D to GENE utilities, which provides
functionality for reading a fluxprof2D file output by the GENE
diagnostic suite.  It can be used to print out the time average of the
fluxes (heat flux or particle flux), which is to be used as the "seed"
in a Tango run.

An example of how to do so is included in example_create_flux_seed.py.