kinetic example modifications
Modified the example kinetic run in order to get it to work.
--Workaround for a bug on quartz where gene would crash in LU
factorization.  Fixed by importing gene_tango before importing other
tango modules (??)
--For the LoDestro method and the convective-diffusive split, added
field-dependent theta parameters,  and a new rule of thumb for
determining these.
--Used a user_control_func that holds the density profile fixed at the
initial condition as we iterate, because it is a tougher problem to let
the density profile evolve.
--other minor things