Minor update for handling restart runs
When performing a Tango restart, AND if the density was fixed on the
previous run, then we need to be careful about setting the initial
density, because what is saved in the last iteration of the hdf5 file is
not that density, but what the density *would have been* had it not been
fixed.

Here, some lines were added that can be used to set the "old profile"
for density, which sets the EWMA iterate.  If restarting from a run
where density was allowed to vary, these lines are not necessary.

Also, a few lines were removed that set the heat/particle flux seeds
(because it is not necessary to do this on a restart---the EWMA iterates
are obtained from the previous iterate).