[lodestro_methpd] Allow a custom ftheta
Allows the user to pass in a custom ftheta to be used instead of the
default ftheta when performing the split of flux into a diffusive and
convective piece.

To reduce the number of changes, a function can be set to the key
'custom_ftheta' in the thetaParams dict.  If this key is set to a
callable, then this function will be used instead of the default.  The
calling signature of the custom function needs to be the same as the
default.