Modifications for Python 3 compatibility
Made modifications for python 3 compatibility.  Old syntax `raise
ValueError, 'txt'` was valid in Python 2 but not Python 3.  new syntax
'raise ValueError('txt')` is valid in both Python 2 and Python 3.

Another weird bug was in the interaction with integer type with h5py.
It has to do with the difference between Python 2 and Python 3 in
treating the int type.  A modification was made that seems to allow it
to work with both Python 2 and 3.