[solver] save profiles for later access
For a given timestep, the solver object will preserve the profiles for the entire iteration history.  This makes it easy to view the profiles after a Tango run.  For example, one could access the profiles using solver.profiles['n'].

For much greater diagnostic info, one could use file output (which is slower), or modify solver.py to save additional data besides the profiles.