Handler option - keep file open
Adding an option to the Tango History Handler to keep the h5 file open each iteration.

This could save time if file open/closing happens to be the bottleneck.  However, in practice, I haven't found that this necessarily made the biggest difference.  File I/O (which is broken up into many separate writes) seems to be the biggest issue, and optimizing that within h5py is more difficult and requires more expertise.