Restructuring flux model decorators
Rather than being classes which implement `get_flux` methods,
now functions which replace the `get_flux` method of given object.

The motivation is to combine the flux calculation method with the
test problem initial and boundary conditions, and exact solution
calculation. The goal is to make testing of different iterations
and boundary condition calculations simpler.